requirements.txt
numpy==1.26.4
scipy==1.11.4
matplotlib==3.8.2
pandas==2.1.4
torch==2.1.2
torchvision==0.16.2
pyyaml==6.0.1
tqdm==4.66.1
